requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
alembic==1.13.2
flask-jwt-extended==4.6.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.25.2